// practice/hw_ang6/src/app/shared/interfaces/settings.interface.ts
/**
 * Export interface ISettings
 */

export interface ISettings {

  /**
   * Variable id
   * @type {string}
   * @memberof ISettings
   */

  id: string;

  /**
   * Variable description
   * @type {string}
   * @memberof ISettings
   */

  description: string;

  /**
   * Variable checked
   * @type {boolean}
   * @memberof ISettings
   */

  checked: boolean;
}